Few more size_t conversions
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@5842 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/simulations/current/apps/tFlux.cc
@@ -88,7 +88,7 @@ int main(int argc, const char** argv)
     double x = 512., y = 512.;
     fluxes.addSpectrum(&cont, x, y, wcs);
 
-    for (int i = 0; i < fluxes.nChan(); i++)
+    for (size_t i = 0; i < fluxes.nChan(); i++)
         std::cout << i << " " << fluxes.getFlux(i) << "\n";
 
     std::cout << "\n";
@@ -97,6 +97,6 @@ int main(int argc, const char** argv)
     cont = Continuum(0., 0., 1.4e9, 1.);
     singleFlux.addSpectrum(&cont, x, y, wcs);
 
-    for (int i = 0; i < singleFlux.nChan(); i++)
+    for (size_t i = 0; i < singleFlux.nChan(); i++)
         std::cout << i << " " << singleFlux.getFlux(i) << "\n";
 }
